Move multiple copies tests to their own file
This moves the express mock and dynamic `require()` loading via mockery
into its own test file.
